Move updateTx
