+++ lib/wallet/src/Cardano/Wallet.hs
@@ -157,7 +157,6 @@ module Cardano.Wallet
     , ErrNotASequentialWallet (..)
     , ErrWithdrawalNotBeneficial (..)
     , ErrConstructTx (..)
-    , ErrUpdateSealedTx (..)
     , ErrCannotJoin (..)
     , ErrCannotQuit (..)
     , ErrSubmitTransaction (..)
@@ -508,7 +507,6 @@ import Cardano.Wallet.Transaction
     , ErrCannotQuit (..)
     , ErrMkTransaction (..)
     , ErrSignTx (..)
-    , ErrUpdateSealedTx (..)
     , PreSelection (..)
     , TransactionCtx (..)
     , TransactionLayer (..)
